setup receptors and start doin some stuff

+++ Average4k/SourceFiles/Average4k/Api/Functions/FGameplay.cpp
@@ -6,7 +6,7 @@
 #include "FGameplay.h"
 #include "../../Screens/Menu/Gameplay.h"
 
-void Average4k::Api::Functions::FGameplay::SetNoteSize(int width, int height)
+void Average4k::Api::Functions::FGameplay::SetNoteSize(float width, float height)
 {
 	Average4k::Screens::Menu::Gameplay* gameplay = (Average4k::Screens::Menu::Gameplay*)A4kGame::gameInstance->CurrentMenu;
