+++ Average4k/SourceFiles/Average4k/Api/Functions/FGameplay.h
@@ -14,7 +14,7 @@ namespace Average4k::Api::Functions
 {
 	class FGameplay {
 	public:
-		static void SetNoteSize(int width, int height);
+		static void SetNoteSize(float width, float height);
 		static void SetComboTag(std::string tag);
 		static void SetJudgementTag(std::string judge, std::string tag);
 		static void SetJudgementTextTag(std::string tag);
